+++ warthog/cli.py
@@ -13,13 +13,67 @@ warthog.cli
 
 CLI interface for interacting with a load balancer using the Warthog client.
 """
+import functools
 
 import os
+
 import os.path
 import click
 import warthog
 import warthog.api
 from . import six
+import requests
+
+
+def error_wrapper(func):
+    """Decorator that coverts possible errors raised by the WarthogClient
+    into instances of ClickExceptions so that they may be rendered automatically
+    """
+
+    # pylint: disable=missing-docstring
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except warthog.api.WarthogNoSuchNodeError as e:
+            raise click.BadParameter("{0} doesn't appear to be a known node".format(e.server))
+        except warthog.api.WarthogAuthFailureError as e:
+            raise click.ClickException(
+                "Authentication with the load balancer failed. The error was: {0}".format(e))
+        except requests.ConnectionError as e:
+            raise click.ClickException(
+                "Connecting to the load balancer failed. The error was {0}".format(e))
+
+    return wrapper
+
+
+class WarthogClientFacade(object):
+    """Wrapper around a :class:`warthog.client.WarthogClient` that coverts
+    exceptions encountered into exceptions that click will handle automatically.
+    """
+
+    def __init__(self, client):
+        self._client = client
+
+    # pylint: disable=missing-docstring
+    @error_wrapper
+    def get_status(self, *args, **kwargs):
+        return self._client.get_status(*args, **kwargs)
+
+    # pylint: disable=missing-docstring
+    @error_wrapper
+    def get_connections(self, *args, **kwargs):
+        return self._client.get_connections(*args, **kwargs)
+
+    # pylint: disable=missing-docstring
+    @error_wrapper
+    def disable_server(self, *args, **kwargs):
+        return self._client.disable_server(*args, **kwargs)
+
+    # pylint: disable=missing-docstring
+    @error_wrapper
+    def enable_server(self, *args, **kwargs):
+        return self._client.enable_server(*args, **kwargs)
 
 
 @click.group()
@@ -46,7 +100,7 @@ def main(ctx, config):
         # BadParameter exceptions with the same message.
         loader.initialize()
     except warthog.api.WarthogConfigError as e:
-        raise click.BadParameter(six.text_type(e))
+        raise click.ClickException(six.text_type(e))
 
     settings = loader.get_settings()
     factory = None
@@ -59,11 +113,13 @@ def main(ctx, config):
         transport = warthog.api.get_transport_factory(verify=False)
         factory = warthog.api.CommandFactory(transport)
 
-    ctx.obj = warthog.api.WarthogClient(
+    # Wrap the client in a facade that translates expected errors into
+    # exceptions that click will render as error messages for the user.
+    ctx.obj = WarthogClientFacade(warthog.api.WarthogClient(
         settings.scheme_host,
         settings.username,
         settings.password,
-        commands=factory)
+        commands=factory))
 
 
 @click.command()
@@ -71,12 +127,9 @@ def main(ctx, config):
 @click.pass_context
 def enable(ctx, server):
     """Enable a server by hostname."""
-    try:
-        if not ctx.obj.enable_server(server):
-            click.echo('{0} could not be enabled'.format(server))
-            ctx.exit(1)
-    except warthog.api.WarthogNoSuchNodeError:
-        raise click.BadParameter("{0} doesn't appear to be a known node".format(server))
+    if not ctx.obj.enable_server(server):
+        click.echo('{0} could not be enabled'.format(server))
+        ctx.exit(1)
 
 
 @click.command()
@@ -84,12 +137,9 @@ def enable(ctx, server):
 @click.pass_context
 def disable(ctx, server):
     """Disable a server by hostname."""
-    try:
-        if not ctx.obj.disable_server(server):
-            click.echo('{0} could not be disabled'.format(server))
-            ctx.exit(1)
-    except warthog.api.WarthogNoSuchNodeError:
-        raise click.BadParameter("{0} doesn't appear to be a known node".format(server))
+    if not ctx.obj.disable_server(server):
+        click.echo('{0} could not be disabled'.format(server))
+        ctx.exit(1)
 
 
 @click.command()
@@ -97,10 +147,7 @@ def disable(ctx, server):
 @click.pass_context
 def status(ctx, server):
     """Get the status of a server by hostname."""
-    try:
-        click.echo(ctx.obj.get_status(server))
-    except warthog.api.WarthogNoSuchNodeError:
-        raise click.BadParameter("{0} doesn't appear to be a known node".format(server))
+    click.echo(ctx.obj.get_status(server))
 
 
 @click.command()
@@ -108,10 +155,7 @@ def status(ctx, server):
 @click.pass_context
 def connections(ctx, server):
     """Get active connections to a server by hostname."""
-    try:
-        click.echo(ctx.obj.get_connections(server))
-    except warthog.api.WarthogNoSuchNodeError:
-        raise click.BadParameter("{0} doesn't appear to be a known node".format(server))
+    click.echo(ctx.obj.get_connections(server))
 
 
 @click.command('default-config')